test: 공고 지원·상세 테스트를 레포 테스트 관례에 맞춰 정리
